+++ src/input.h
@@ -3,8 +3,8 @@
 #define _INPUT_H_
 
 struct input_handle;
-int input_get_channels(struct input_handle* ih);
-int input_get_samplerate(struct input_handle* ih);
+size_t input_get_channels(struct input_handle* ih);
+size_t input_get_samplerate(struct input_handle* ih);
 float* input_get_buffer(struct input_handle* ih);
 size_t input_get_buffer_size(struct input_handle* ih);
 struct input_handle* input_handle_init();
